Fix code that leaves and joins multicast group.
There is code to leave and join again the multicast group if no network
packets are received within a given timeout. This code was wrongly
updating the "last received time" also when only checking for the
timeout.

This code was fixed, and the variable renamed to better clarify its
intention.

Signed-off-by: Rafael Antognolli <rafael.antognolli@intel.com>