Remove time_to_send increment on pause.
This logic was added to handle pause on the stream sharing, but since we
are not using pause right now, it's better to just remove this code to
reduce the complexity on the code that handles network feedback.

This code was also causing the network timeout to not work correctly,
since it was never calling the CheckNetworkTimeout() function.

Signed-off-by: Rafael Antognolli <rafael.antognolli@intel.com>