Set max capture size to 4096x4096.
Setting it to this high value will mean pratically no restriction on the
video size (4k would probably be the maximum size). The size of the
frames on the stream will be the size of the original screen.

We can then later adjust the size of encoding within the encoder itself,
where we will have information about the frames size.

Signed-off-by: Rafael Antognolli <rafael.antognolli@intel.com>